fix: reject invalidated edges as projection inputs, guard NULL anchor uniqueness, split types
